requirements.txt
Click==7.0
Flask==1.0.2
imutils==0.5.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.16.3
opencv-contrib-python==3.4.2.16
opencv-python==3.4.2.16
pdf2image==1.5.4
Pillow==6.0.0
pytesseract==0.2.6
Werkzeug==0.15.2
